+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- concord-project-model: `exit` step to terminate execution of
+the flow w/o throwing an error;
 - concord-client: support for the new `startAt` parameter.
 
 ### Changed